升级dili-parent父版本为3.7.0；升级项目版本为1.2.8 接入sentinel监控

+++ customer/src/main/java/com/dili/customer/service/impl/CustomerServiceImpl.java
@@ -213,7 +213,7 @@ public class CustomerServiceImpl extends BaseServiceImpl<Customer, Long> impleme
                 });
             }
         }
-        output.setData(list).setPageNum(pageNum).setTotal(total.intValue()).setPageSize(input.getPage()).setPages(totalPage);
+        output.setData(list).setPageNum(pageNum).setTotal(total).setPageSize(input.getPage()).setPages(totalPage);
         return output;
     }
 
@@ -236,7 +236,7 @@ public class CustomerServiceImpl extends BaseServiceImpl<Customer, Long> impleme
         //当前页数
         int pageNum = list instanceof Page ? ((Page) list).getPageNum() : 1;
         PageOutput output = PageOutput.success();
-        output.setData(list).setPageNum(pageNum).setTotal(total.intValue()).setPageSize(input.getPage()).setPages(totalPage);
+        output.setData(list).setPageNum(pageNum).setTotal(total).setPageSize(input.getPage()).setPages(totalPage);
         return output;
     }
